Krp development

See merge request mra.narola/clinic-app!40

+++ app/Http/Controllers/Doctor/DoctorController.php
@@ -593,7 +593,8 @@ class DoctorController extends Controller
             })
             ->addColumn('appointment_date',function($row){
                 if ($row->next_date) {
-                    return date('d-m-Y H:i a',strtotime($row->next_date .' '.$row->next_start_time));    
+                    $next_date = date("Y-m-d",strtotime($row->next_date));
+                    return date('d-m-Y H:i a',strtotime($next_date .' '.$row->next_start_time));    
                 }
                 return date('d-m-Y H:i a',strtotime($row->appointment_date.' '.$row->time_start));
             })
@@ -727,7 +728,8 @@ class DoctorController extends Controller
             })
             ->addColumn('appointment_date',function($row){
                 if ($row->next_date) {
-                    return date('d-m-Y H:i a',strtotime($row->next_date .' '.$row->next_start_time));    
+                    $next_date = date("Y-m-d",strtotime($row->next_date));
+                    return date('d-m-Y H:i a',strtotime($next_date .' '.$row->next_start_time));     
                 }
                 return date('d-m-Y H:i a',strtotime($row->appointment_date.' '.$row->time_start));
             })
@@ -854,7 +856,8 @@ class DoctorController extends Controller
             })
             ->addColumn('appointment_date',function($row){
                 if ($row->next_date) {
-                    return date('d-m-Y H:i a',strtotime($row->next_date .' '.$row->next_start_time));    
+                    $next_date = date("Y-m-d",strtotime($row->next_date));
+                    return date('d-m-Y H:i a',strtotime($next_date .' '.$row->next_start_time));    
                 }
                 return date('d-m-Y H:i a',strtotime($row->appointment_date.' '.$row->time_start));
             })
@@ -878,6 +881,9 @@ class DoctorController extends Controller
                 return date('H:i A', strtotime($row->time_end));
             })
             ->addColumn('action', function($row){
+                if($row->deleted_at){
+                    return '-';
+                }
                  $actionBtn =   '<div class="dropable-btn">
                                     <div class="dropdown">
                                         <button class="btn dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">
@@ -1091,14 +1097,13 @@ class DoctorController extends Controller
             'patient_id' => $user_id,
         ))->get();
 
-
         $available_slot[] = '';
         foreach ($available_time_slot as $available) {
             $available_slot[] = $available->next_start_time . ' - ' . $available->next_end_time;
         }   
 
         $general_time = GeneralSettings::select('start_time', 'end_time', 'duration')->where('user_id', $user_id)->first();
-
+        
         $current_date = date('H:00:00');
         
         $current_time = now()->toTimeString();
@@ -1134,10 +1139,7 @@ class DoctorController extends Controller
                 $time[$i]['end'] = $end;
             }
             $i++;
-    
         } 
-        
-       
        
         $all_appointent = DoctorAppointmentDetails::with('patient')->findOrFail($id);        
         $appointment_history=DoctorAppointmentDetails::select('id','user_id','doctor_id','appointment_date','time_start','time_end'
@@ -1150,6 +1152,7 @@ class DoctorController extends Controller
                 'view' => $view,
             ),
         );
+        
         return response()->json($response);
     }
 
@@ -1739,6 +1742,9 @@ class DoctorController extends Controller
                 return date('H:i A', strtotime($row->time_end));
             })
             ->addColumn('action', function($row){
+                if($row->deleted_at){
+                    return '-';
+                }
                  $actionBtn =   '<div class="dropable-btn">
                                            <a class="dropdown-item" href="javascript:delete_record(' . $row->id . ');" class="delete btn btn-delete" title="Delete">
                                                   <span class="svg-icon">
